bugfix: autoframes mode
when frames < 0 and pad >= 0 the behavior is such that once the txbuff
is found to be empty, Py_PaStream will output zeros for ``pad`` more
frames; this is the 'autoframes' mode. This commit makes sure that once
autoframes mode is entered, Py_PaStream will not attempt to read from
the txbuff again.